requirements.txt
setuptools
tqdm
nmslib
matplotlib
scikit-learn
Pillow
imgaug
imutils
keras==2.2.4
tensorflow==1.6
numpy==1.15.4
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16